feat: split auth page
