+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 
 import MainLayout from './layouts/MainLayout';
-import { Auth, Main, NotFound, Welcome } from 'pages';
+import { Login, Register, Main, NotFound, Welcome } from 'pages';
 import { Paths } from 'assets';
 
 const App = () => {
@@ -9,7 +9,8 @@ const App = () => {
     <Routes>
       <Route path={Paths.WELCOME} element={<MainLayout />}>
         <Route index element={<Welcome />} />
-        <Route path={Paths.AUTH} element={<Auth />} />
+        <Route path={Paths.AUTH} element={<Login />} />
+        <Route path={Paths.SIGNUP} element={<Register />} />
         <Route path={Paths.MAIN} element={<Main />} />
       </Route>
       <Route path="*" element={<NotFound />} />
