--- src/assets/authOptions.ts
@@ -1,11 +1,0 @@
-export const LoginDefinition = {
-  title: 'Sign in',
-  text: "Don't have an account? ",
-  linkText: 'Register here',
-};
-
-export const RegisterDefinition = {
-  title: 'Sign up',
-  text: 'Have an account? ',
-  linkText: 'Sign in',
-};
