+++ src/assets/index.ts
@@ -1,3 +1,2 @@
-export * from './authOptions';
 export * from './editorOptions';
 export * from './routes';
