+++ src/components/AuthFields/AuthFields.module.scss
@@ -1,0 +1,3 @@
+.error {
+  color: #ff0000;
+}
