+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export { Footer } from './Footer';
 export { Header } from './Header';
 export { Editors } from './Editors';
-export { AuthForm } from './AuthForm';
 export { Documentation } from './Documentation';
 export { AuthCheck } from './AuthCheck';
+export * from './AuthFields';
