+++ src/pages/Login/Login.module.scss
@@ -23,7 +23,3 @@
     color: #ff0000;
   }
 }
-
-.error {
-  color: #ff0000;
-}
