+++ src/pages/index.tsx
@@ -1,4 +1,5 @@
-export { Auth } from './Auth';
+export { Login } from './Login';
+export { Register } from './Register';
 export { Main } from './Main';
 export { NotFound } from './NotFound';
 export { Welcome } from './Welcome';
