+++ src/types/auth.ts
@@ -3,9 +3,3 @@ export type Inputs = {
   email: string;
   password: string;
 };
-
-export interface FormDefinition {
-  title: string;
-  text: string;
-  linkText: string;
-}
